changes for improving performance of XMTP context
