+++ components/NavigationPanel.tsx
@@ -1,6 +1,6 @@
 import { LinkIcon } from '@heroicons/react/outline'
-import { ChatIcon } from '@heroicons/react/outline'
 import { ArrowSmRightIcon } from '@heroicons/react/solid'
+import useWallet from '../hooks/useWallet'
 import useXmtp from '../hooks/useXmtp'
 import ConversationsList from './ConversationsList'
 import Loader from './Loader'
@@ -14,7 +14,7 @@ type ConnectButtonProps = {
 }
 
 const NavigationPanel = ({ onConnect }: NavigationPanelProps): JSX.Element => {
-  const { walletAddress } = useXmtp()
+  const { address: walletAddress } = useWallet()
 
   return (
     <div className="flex-grow flex flex-col h-[82vh] overflow-y-auto">
@@ -64,7 +64,8 @@ const ConnectButton = ({ onConnect }: ConnectButtonProps): JSX.Element => {
 }
 
 const ConversationsPanel = (): JSX.Element => {
-  const { conversations, loadingConversations, client } = useXmtp()
+  const { loadingConversations, client } = useXmtp()
+
   if (!client) {
     return (
       <Loader
@@ -74,6 +75,7 @@ const ConversationsPanel = (): JSX.Element => {
       />
     )
   }
+
   if (loadingConversations) {
     return (
       <Loader
@@ -84,31 +86,10 @@ const ConversationsPanel = (): JSX.Element => {
     )
   }
 
-  return conversations && conversations.length > 0 ? (
-    <nav className="flex-1 pb-4 space-y-1">
-      <ConversationsList conversations={conversations} />
-    </nav>
-  ) : (
-    <NoConversationsMessage />
-  )
-}
-
-const NoConversationsMessage = (): JSX.Element => {
   return (
-    <div className="flex flex-col flex-grow justify-center">
-      <div className="flex flex-col items-center px-4 text-center">
-        <ChatIcon
-          className="h-8 w-8 mb-1 stroke-n-200 md:stroke-n-300"
-          aria-hidden="true"
-        />
-        <p className="text-xl md:text-lg text-n-200 md:text-n-300 font-bold">
-          Your message list is empty
-        </p>
-        <p className="text-lx md:text-md text-n-200 font-normal">
-          There are no messages in this wallet
-        </p>
-      </div>
-    </div>
+    <nav className="flex-1 pb-4 space-y-1">
+      <ConversationsList />
+    </nav>
   )
 }
