+++ contexts/xmtp.ts
@@ -10,7 +10,6 @@ export type MessageStoreEvent = {
 
 export type XmtpContextType = {
   wallet: Signer | undefined
-  walletAddress: string | undefined
   client: Client | undefined
   conversations: Conversation[]
   loadingConversations: boolean
@@ -22,7 +21,6 @@ export type XmtpContextType = {
 
 export const XmtpContext = createContext<XmtpContextType>({
   wallet: undefined,
-  walletAddress: undefined,
   client: undefined,
   conversations: [],
   loadingConversations: false,
